refactor(models): extract VerdictValue to models layer to break circular dependency
Batch 1A (partial) of modular refactoring.

Changes:
- Create models/verdict_types.py with VerdictValue type definition
- Update services/verdict_policy.py to import from models.verdict_types
- Update models/verdict.py to import from models.verdict_types
- Update 4 other files to import VerdictValue from models layer
  - services/verdict_service.py
  - services/flow_resume_resolver.py
  - commands/handoff_write.py
  - execution/noop_gate.py

Verification:
- All verdict tests pass (30 tests)
- models/verdict.py no longer imports from services
- VerdictValue is now a pure type in models layer

Remaining work:
- models/orchestra_config.py still re-exports from config layer
- Need to move actual config classes to models layer
- This affects 30+ files and requires careful migration
