+++ src/vibe3/services/verdict_service.py
@@ -14,11 +14,11 @@ from loguru import logger
 from vibe3.clients import SQLiteClient
 from vibe3.clients.git_client import GitClient
 from vibe3.models.verdict import VerdictRecord
+from vibe3.models.verdict_types import VerdictValue
 from vibe3.services.actor_support import extract_role_from_actor
 from vibe3.services.flow_service import FlowService
 from vibe3.services.handoff_storage import HandoffStorage
 from vibe3.services.signature_service import SignatureService
-from vibe3.services.verdict_policy import VerdictValue
 from vibe3.utils.git_path_client import GitPathProtocol
